apply shell script condition

+++ roles/deploy/files/stop.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
 pid=`ps -ef | grep sp-0.0.1-SNAPSHOT.jar | grep -v 'grep' | awk '{print $2}'`
-echo $pid
-kill -9 $pid
+if [ -n "$pid" ]; then
+  echo $pid
+  kill -9 $pid
+fi
